requirements.txt
# Python 3.10+
python-telegram-bot==20.7
openai==1.3.5
python-dotenv==1.0.0